didn't save before commit

+++ Cell.java
@@ -1,2 +1,2 @@
-public record Cell(int x, int y) {
+public record Cell(long x, long y) {
 }
